refactor: extract common index references to new module
