+++ migration/commons/src/main/java/io/camunda/migration/commons/storage/ProcessorStep.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.migration.process.adapter;
+package io.camunda.migration.commons.storage;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
