+++ migration/commons/src/main/java/io/camunda/migration/commons/storage/TasklistMigrationRepositoryIndex.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.migration.commons.storage;
+
+import io.camunda.webapps.schema.descriptors.ComponentNames;
+
+public class TasklistMigrationRepositoryIndex extends MigrationRepositoryIndex {
+
+  public TasklistMigrationRepositoryIndex(final String indexPrefix, final boolean isElasticsearch) {
+    super(indexPrefix, isElasticsearch);
+  }
+
+  @Override
+  public String getComponentName() {
+    return ComponentNames.TASK_LIST.toString();
+  }
+}
