+++ migration/process-migration/src/main/java/io/camunda/migration/process/adapter/Adapter.java
@@ -8,6 +8,7 @@
 package io.camunda.migration.process.adapter;
 
 import io.camunda.migration.api.MigrationException;
+import io.camunda.migration.commons.storage.ProcessorStep;
 import io.camunda.webapps.schema.descriptors.index.ProcessIndex;
 import io.camunda.webapps.schema.entities.ImportPositionEntity;
 import io.camunda.webapps.schema.entities.ProcessEntity;
