+++ migration/process-migration/src/main/java/io/camunda/migration/process/adapter/os/OpensearchAdapter.java
@@ -8,9 +8,9 @@
 package io.camunda.migration.process.adapter.os;
 
 import io.camunda.migration.api.MigrationException;
+import io.camunda.migration.commons.storage.MigrationRepositoryIndex;
+import io.camunda.migration.commons.storage.ProcessorStep;
 import io.camunda.migration.process.adapter.Adapter;
-import io.camunda.migration.process.adapter.MigrationRepositoryIndex;
-import io.camunda.migration.process.adapter.ProcessorStep;
 import io.camunda.migration.process.config.ProcessMigrationProperties;
 import io.camunda.search.connect.configuration.ConnectConfiguration;
 import io.camunda.search.connect.os.OpensearchConnector;
