+++ parent/pom.xml
@@ -2173,6 +2173,12 @@
 
       <dependency>
         <groupId>io.camunda.migration</groupId>
+        <artifactId>migration-commons</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>io.camunda.migration</groupId>
         <artifactId>process-migration</artifactId>
         <version>${project.version}</version>
       </dependency>
